% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_AFSCforeign_hauls.R
\docType{data}
\name{AFSCforeign_hauls}
\alias{AFSCforeign_hauls}
\title{AFSC Haul join characteristics for 409 hauls.}
\format{A data frame with 409 rows and 2 variables:
\describe{
\item{HAULJOIN}{unique row ID}
\item{Jurisdctn}{jurisdiction id}
...
}}
\usage{
AFSCforeign_hauls
}
\description{
An example dataset for the nwfscDeltaGLM package.
}
\keyword{datasets}
