\name{cMx}
\alias{cMx}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Converts and object to a matrix}
\description{
  Converts and object to a matrix.
}
\usage{
cMx(Input,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Input}{A scalar, data frame, etc that is converted to a returned matrix object}
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
