\name{ComputeIndices}
\alias{ComputeIndices}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Computes indices from fitted delta-GLMM model}
\description{
  }
\usage{
ComputeIndices(Data, Model, FileName, Folder=NA, Weights="StrataAreas", StrataTable, Run=TRUE,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{}
  \item{Model}{}
  \item{FileName}{}
  \item{Folder}{}
  \item{Weights}{}  
  \item{StrataTable}{}  
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
