\name{ConvergencePlot}
\alias{ConvergencePlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Makes plots related to MCMC convergence diagnostics}
\description{
}
\usage{
ConvergencePlot(McmcArray, maxDims=8, parToMonitor, parnames, Nkeep=5000, FileName, Type="Trace", Folder=NA, Model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{McmcArray}{}
  \item{maxDims}{}
  \item{parToMonitor}{}
  \item{parnames}{}
  \item{Nkeep}{}
  \item{FileName}{}
  \item{Type}{}
  \item{Folder}{}
  \item{Model}{}  
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
