% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFunction.r
\name{corFunction}
\alias{corFunction}
\title{Function to calculate the correlations from the precision matrix}
\usage{
corFunction(McmcArray, Parameter, this.names, Model)
}
\arguments{
\item{McmcArray}{The array of MCMC output}

\item{Parameter}{The names of all parameters monitoried}

\item{this.names}{The name of the parameter of interest}

\item{Model}{The fitted model object}
}
\value{
Returns a list of output, by year and by Strata:Year
}
\description{
Function to calculate the correlations from the precision matrix
}
