% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_defaultLimits.R
\docType{data}
\name{defaultLimits}
\alias{defaultLimits}
\title{Haul characteristics for 1756 hauls.}
\format{
A data frame with 1756 rows and 7 variables:
\describe{
\item{STRATA}{strata name}
\item{NLat}{northern boundary, in decimal degrees}
\item{SLat}{southern boundary, in decimal degrees}
\item{MinDepth}{minimum depth, in meters}
\item{MaxDepth}{maximum depth, in meters}
...
}
}
\usage{
defaultLimits
}
\description{
An example dataset for the nwfscDeltaGLM package.
}
\keyword{datasets}
