% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMCMCDiags.r
\name{doMCMCDiags}
\alias{doMCMCDiags}
\title{This function is a high level function to be called by the user, and implements all of the plotting, diagnostic, and output summaries available. It is applicable to one or more fitted model objects, stored in a list object.}
\usage{
doMCMCDiags(datalist, strata.limits, directory, mods,
  StrataWeights = "StrataAreas", McmcDiagnostics = FALSE)
}
\arguments{
\item{datalist}{Data}

\item{strata.limits}{limits of strata}

\item{directory}{The directory to save output in}

\item{mods}{A list of one or more fitted delta-GLMM model objects}

\item{StrataWeights}{Weights, defaults to "StrataAreas", also can be "Equal"}

\item{McmcDiagnostics}{Boolean indicating whether or not to make the MCMCdiagnostics plots, defaults to \code{FALSE}}
}
\value{
Returns a list of output, by year and by Strata:Year
}
\description{
This function is a high level function to be called by the user, and implements all of the plotting, diagnostic, and output summaries available. It is applicable to one or more fitted model objects, stored in a list object.
}
\seealso{
\code{\link{MapData}}, \code{\link{ConvergencePlot}}, \code{\link{McmcDiagnosticsPlot}}, \code{\link{PlotOffset}}, \code{\link{PosteriorPredictive}}, \code{\link{ComputeIndices}}, \code{\link{ComputeMleIndices}}
}
