% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFn.r
\name{LoadFn}
\alias{LoadFn}
\title{Load functions}
\usage{
LoadFn(file, ...)
}
\arguments{
\item{file}{file name to load}

\item{\\dots}{additional arguments
\describe{
\item{directory}{The directory to save output in}
\item{mods}{A list of one or more fitted delta-GLMM model objects}
\item{McmcDiagnostics}{Boolean indicating whether or not to make the MCMCdiagnostics plots, defaults to FALSE.}
}}
}
\description{
Load functions
}
