% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDensity.R
\name{logDensity}
\alias{logDensity}
\title{This function is in development stage, but used to generate the posterior predictive distribution score, or log score, from a fitted model object. Currently, it's only applicable to the gamma, lognormal, or inverse Gaussian likelihoods.}
\usage{
logDensity(obj)
}
\arguments{
\item{obj}{A fitted modle object}
}
\value{
avg.bin The log score for the presence-absence model
}
\description{
This function is in development stage, but used to generate the posterior predictive distribution score, or log score, from a fitted model object. Currently, it's only applicable to the gamma, lognormal, or inverse Gaussian likelihoods.
}
