\name{PlotData}
\alias{PlotData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{PlotData is the function that makes plots of raw data before without fitting statistical models}
\description{Makes the following plots: (1) Histogram of positive catch by year, (2) Scatterplot of positive catch by depth, (3) Scatterplot of positive catch by depth by year, (4) Scatterplot of positive catch by lattitude, (5) Scatterplot of positive catch by lattitude by year, (6) bar graph of proportion positive by year, (7) bar graph of proportion positive by 25 meter depth bins by year, (7) bar graph of proportion positive by 1 degree latitude bins by year
}
\usage{
PlotData(Data, FileName, Folder=NA,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{The data frame with the raw survey data}
  \item{FileName}{A species and model specific name created for output}
  \item{Folder}{Where to put the output}
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
