\name{PlotOffset}
\alias{PlotOffset}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{PlotOffset plots the positive and presence absence offset}
\description{Makes plots of the positive and presence-absence model by strata}
\usage{
PlotOffset(Data, BugsList, maxDims=8, FileName, Folder=NA,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{The data frame with the raw survey data}
  \item{BugsList}{List of the fitted BUGS / JAGS object}
  \item{maxDims}{Max dimensions}
  \item{FileName}{Name of the file(s) created}
  \item{Folder}{Where to put the output is to be stored}  
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
