% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorPredictive.r
\name{PosteriorPredictive}
\alias{PosteriorPredictive}
\title{Experimental function for calculating log-scores}
\usage{
PosteriorPredictive(Data, Model, maxDims = 6, FileName, Folder = NA)
}
\arguments{
\item{Data}{data used to fit}

\item{Model}{Fitted model object}

\item{maxDims}{Optional argument for specifying grid size, defaults to 6}

\item{FileName}{Used for output}

\item{Folder}{Used for output}
}
\value{
data frame of data
}
\description{
Experimental function for calculating log-scores
}
