% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.r
\name{processData}
\alias{processData}
\title{Process raw haul data}
\usage{
processData(Truncate = 0)
}
\arguments{
\item{Truncate}{Lower bound of hauls, defaults to 0. Values < Truncate will be turned to 0}
}
\value{
DataList Data frame of processed data to fit
}
\description{
Process raw haul data
}
