% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIn.R
\name{readIn}
\alias{readIn}
\title{Scan data}
\usage{
readIn(ncol, nlines, ...)
}
\arguments{
\item{ncol}{number of columns}

\item{nlines}{Number of lines}

\item{\\dots}{Additional arguments passed to the function}
}
\value{
data frame of data
}
\description{
Scan data
}
