% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SA3.R
\docType{data}
\name{SA3}
\alias{SA3}
\title{Haul characteristics for 1756 hauls.}
\format{
A data frame with 1756 rows and 7 variables:
\describe{
\item{SUBAREA_ID}{area ID}
\item{MIN_DEPTH_M}{minimum haul depth}
\item{MAX_DEPTH_M}{maximum haul depth}
\item{AREA_HECTARES}{area swept, proxy for effort, in hectares}
\item{SUBAREA_SET_ID}{set id}
\item{MIN_LAT_DD}{minimum latitude}
\item{MAX_LAT_DD}{maximum latitude}
...
}
}
\usage{
SA3
}
\description{
An example dataset for the nwfscDeltaGLM package.
}
\keyword{datasets}
