% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSimple.R
\name{SimSimple}
\alias{SimSimple}
\title{Simulate data}
\usage{
SimSimple(MeanEncounter, Nyears = 10, Nstrata = 15, Nvessels = 4,
  Obsperyear = 175, sigmaV = rep(1, 2), sigmaVY = rep(1, 2),
  sigmaS = rep(1, 2), sigmaY = rep(1, 2), sigmaSY = rep(1, 2),
  sigmaResid = 1)
}
\arguments{
\item{MeanEncounter}{mean encounter rate}

\item{Nyears}{Number of years, defaults to 10}

\item{Nstrata}{Number of strata, defaults to 15}

\item{Nvessels}{Number of vessels, defaults to 4}

\item{Obsperyear}{Data points per year, defaults to 175}

\item{sigmaV}{Two element vector of standard deviations of vessel random effects, defaults to \code{c(1,1)}}

\item{sigmaVY}{Two element vector of standard deviations of vessel-year random effects, defaults to \code{c(1,1)}}

\item{sigmaS}{Two element vector of standard deviations of strata random effects, defaults to \code{c(1,1)}}

\item{sigmaY}{Two element vector of standard deviations of year random effects, defaults to \code{c(1,1)}}

\item{sigmaSY}{Two element vector of standard deviations of strata-year random effects, defaults to \code{c(1,1)}}

\item{sigmaResid}{Residual sd, defaults to \code{1}}
}
\value{
List of simulated data
}
\description{
Simulate data
}
