% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.fn.r
\name{strata.fn}
\alias{strata.fn}
\title{Assign latitude and longitude to strata}
\usage{
strata.fn(x, Strata.df)
}
\arguments{
\item{x}{The raw data, with latitude in \code{BEST_LAT_DD} and longitude in \code{BEST_LON_DD} and depth in \code{BEST_DEPTH_M}}

\item{Strata.df}{Data frame of strata boundaries}
}
\value{
Char Character vector of strate assignments
}
\description{
Assign latitude and longitude to strata
}
