% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeIndices.R
\name{ComputeIndices}
\alias{ComputeIndices}
\title{Compute index of abundance}
\usage{
ComputeIndices(
  Data,
  Model,
  FileName,
  maxDims = 6,
  Folder = NA,
  Weights = "StrataAreas",
  StrataTable,
  PlotStrataYearMcmc = TRUE
)
}
\arguments{
\item{Data}{The data to be fit}

\item{Model}{The model object (as a list)}

\item{FileName}{The name of the file/directory}

\item{maxDims}{aesthetics argument for plot as dimensions of rows/columns, defaults to 6}

\item{Folder}{Optional argument for named folder to store results (in current working directory)}

\item{Weights}{Defaults to "StrataAreas", other option is "Equal"}

\item{StrataTable}{Table of strata, required}

\item{PlotStrataYearMcmc}{Whether to plot MCMC chains for each Strata:Year, defaults to \code{TRUE}}
}
\value{
Returns a list of output, by year and by Strata:Year
}
\description{
Compute index of abundance
}
