% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMleIndices.R
\name{ComputeMleIndices}
\alias{ComputeMleIndices}
\title{Compute MLE index of abundance}
\usage{
ComputeMleIndices(Data, Model, FileName, Folder = NA,
  Weights = "StrataAreas", StrataTable, Run = TRUE)
}
\arguments{
\item{Data}{The data to be fit}

\item{Model}{The model object (as a list)}

\item{FileName}{The name of the file/directory}

\item{Folder}{Optional argument for named folder to store results (in current working directory)}

\item{Weights}{Defaults to "StrataAreas", other option is "Equal"}

\item{StrataTable}{Table of strata, required}

\item{Run}{Defaults to \code{TRUE}. Only run if there's no random effects and the distribution is either Gamma and Lognormal}
}
\value{
Returns a list of output, by year and by Strata:Year
}
\description{
Compute MLE index of abundance
}
