% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvergencePlot.r
\name{ConvergencePlot}
\alias{ConvergencePlot}
\title{Compute MLE index of abundance}
\usage{
ConvergencePlot(
  McmcArray,
  maxDims = 8,
  parToMonitor,
  parnames,
  Nkeep = 5000,
  FileName,
  Type = "Trace",
  Folder = NA,
  Model
)
}
\arguments{
\item{McmcArray}{The array of MCMC output}

\item{maxDims}{aesthetics argument for plot (dimensions of rows/columns), defaults to 8}

\item{parToMonitor}{The parameters to be monitored}

\item{parnames}{The names of parameters to be plotted}

\item{Nkeep}{Number of parameter sets to save, defaults to 5000}

\item{FileName}{The name of the file/directory}

\item{Type}{Type of plot - defaults to "Trace", alternative is "ACF"}

\item{Folder}{Optional argument for named folder to store results (in current working directory)}

\item{Model}{The fitted model object}
}
\value{
Returns a list of output, by year and by Strata:Year
}
\description{
Compute MLE index of abundance
}
