\name{corFunction}
\alias{corFunction}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to calculate the correlations from the precision matrix}
\description{
}
\usage{
corFunction(McmcArray, Parameter,this.names, Model,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{McmcArray}{}
  \item{Parameter}{}
  \item{this.names}{}
  \item{Model}{}
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
