\name{loadFn}
\alias{loadFn}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{loadFn}
\description{}
\usage{
loadFn(directory, mods, McmcDiagnostics=FALSE,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{directory}{The directory to save output in}
  \item{mods}{A list of one or more fitted delta-GLMM model objects}
  \item{McmcDiagnostics}{Boolean indicating whether or not to make the MCMCdiagnostics plots, defaults to FALSE.}
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
