\name{logDensity}
\alias{logDensity}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{logDensity}
\description{This function is in development stage, but used to generate the posterior predictive distribution score, or log score, from a fitted model object. Currently, it's only applicable to the gamma, lognormal, or inverse Gaussian likelihoods.
}
\usage{
logDensity(obj,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{The directory to save output in}
  \item{\dots}{Other arguments can be passed.}  
}
\value{
\item{avg.bin}{The log score for the presence-absence model}
}
\seealso{}
