% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapData.r
\name{MapData}
\alias{MapData}
\title{Plot tows on a map. This function produces a plot (originally designed by John Wallace) showing tows in strata by depth and latitude. If a folder name is given a plot called TowMap.jpg is created.}
\usage{
MapData(Data, strata.limits, SA = SA3, FileName = "TowMap.jpg", Folder = NULL)
}
\arguments{
\item{Data}{The data frame storing raw tow data}

\item{strata.limits}{A file (e.g. defaultLimits.csv), indicating an alphabetic character for each strata (A, B, C, ...), as well as the depth and latitude ranges for each stratum.}

\item{SA}{A flat data file, with area ID, min / max depth, min / max latitude, and the area of each strata in hectares.}

\item{FileName}{The name of the output file -- this will be species and model specific. Defaults to 'TowMap.jpg'}

\item{Folder}{Where the output is stored}
}
\description{
Plot tows on a map. This function produces a plot (originally designed by John Wallace) showing tows in strata by depth and latitude. If a folder name is given a plot called TowMap.jpg is created.
}
