% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCDiagnosticsPlot.R
\name{McmcDiagnosticsPlot}
\alias{McmcDiagnosticsPlot}
\title{Gelman-Rubin diagnostics. This function does MCMC diagnostics, producing several plots related to Gelman diagnostics, and optionally a plot for the Geweke statistics (though this doesn't converge sometimes)}
\usage{
McmcDiagnosticsPlot(
  McmcList,
  parToMonitor,
  FileName = "",
  Folder = NA,
  Geweke = FALSE
)
}
\arguments{
\item{McmcList}{An MCMC list object of model parameters}

\item{parToMonitor}{Which parameter(s) of the fitted model object were monitored / estimated}

\item{FileName}{Species and model specific, where the outupt will be generated to}

\item{Folder}{Folder for output, defaults to working directory}

\item{Geweke}{Boolean, whether to also calculate Geweke diagnostics, defaults to FALSE}
}
\description{
Gelman-Rubin diagnostics. This function does MCMC diagnostics, producing several plots related to Gelman diagnostics, and optionally a plot for the Geweke statistics (though this doesn't converge sometimes)
}
