% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotData.R
\name{PlotData}
\alias{PlotData}
\title{Plot data}
\usage{
PlotData(Data, FileName, Folder = NA)
}
\arguments{
\item{Data}{The data frame storing raw tow data}

\item{FileName}{The name of the output file -- this will be species and model specific. Defaults to 'TowMap.jpg'}

\item{Folder}{Where the output is stored, defaults to NA}
}
\description{
Plot data
}
