% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOffset.r
\name{PlotOffset}
\alias{PlotOffset}
\title{PlotOffset plots the positive and presence absence offset by strata}
\usage{
PlotOffset(Data, BugsList, maxDims = 8, FileName, Folder = NA)
}
\arguments{
\item{Data}{The data frame with the raw survey data}

\item{BugsList}{List of the fitted BUGS / JAGS object}

\item{maxDims}{Max dimensions of grid (rows, columns). Defaults to 8}

\item{FileName}{Name of the file(s) created}

\item{Folder}{Where to put the output is to be stored, defaults to NA}
}
\description{
PlotOffset plots the positive and presence absence offset by strata
}
