\name{PosteriorPredictive}
\alias{PosteriorPredictive}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{PosteriorPredictive is the function that converts the raw trawl data file into a format that is used by the JAGS model}
\description{PosteriorPredictive attaches the appropriate variables to the global workspace}
\usage{
PosteriorPredictive(Data, Model, maxDims=6, FileName, Folder=NA,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{}
  \item{Model}{}
  \item{maxDims}{} 
  \item{FileName}{}  
  \item{Folder}{}    
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
