\name{readIn}
\alias{readIn}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{readIn...}
\description{reads in columns / lines from the command window / screen; written by J. Wallace}
\usage{
readIn(ncol, nlines,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ncol}{Number of columns to read in}
  \item{nlines}{Number of lines to read in} 
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
