\name{SimSimple}
\alias{SimSimple}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{SimSimple...}
\description{Simple simulator for data to test the delta-GLMM estimator}
\usage{
Sim_Simple(MeanEncounter,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MeanEncounter}{Scalar between 0 and 1, representing the median catch rate, which can be used to define either a data-rich or data-poor species}
  \item{\dots}{Other arguments can be passed.}
}
\value{}
\seealso{}
