\name{strata.fn}
\alias{strata.fn}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{strata.fn...}
\description{Assigns observations / locations to strata, based on function by A. Hicks}
\usage{
strata.fn(x,Strata.df,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Raw observations}
  \item{Strata.df}{Data frame containing strata limits} 
  \item{\dots}{Other arguments can be passed.}  
}
\value{}
\seealso{}
